+++ ffn_dl/erotica/literotica.py
@@ -18,7 +18,7 @@ import re
 from bs4 import BeautifulSoup
 
 from ..models import Chapter, Story, chapter_in_spec
-from ..scraper import BaseScraper, StoryNotFoundError
+from ..scraper import BaseScraper
 
 logger = logging.getLogger(__name__)