+++ ffn_dl/wattpad.py
@@ -29,7 +29,6 @@ import json
 import logging
 import re
 
-from bs4 import BeautifulSoup
 
 from .models import Chapter, Story, chapter_in_spec
 from .scraper import BaseScraper, StoryNotFoundError